+++ src/cpp/src/nn/layers/embedding/embedding.cpp
@@ -16,7 +16,7 @@ EmbeddingLayer::EmbeddingLayer(shared_ptr<LayerConfig> layer_config, torch::Devi
 
 torch::Tensor EmbeddingLayer::forward(torch::Tensor input) { return input.narrow(1, offset_, config_->output_dim); }
 
-torch::Tensor EmbeddingLayer::init_embeddings(int64_t num_nodes) {
+torch::Tensor EmbeddingLayer::init_embeddings(int32_t num_nodes) {
     auto options = torch::TensorOptions().device(device_).dtype(torch::kFloat32);
     torch::Tensor embs = initialize_tensor(config_->init, {num_nodes, config_->output_dim}, options);
